# Make build compatible of -DCMAKE_UNITY_BUILD=ON for faster builds