--- a/src/conversions/axisswap.cpp
+++ b/src/conversions/axisswap.cpp
@@ -52,7 +52,6 @@ operation:
 *
 ***********************************************************************/
 
-#define PJ_LIB_
 #include <errno.h>
 #include <stdlib.h>
 #include <string.h>