--- a/src/conversions/cart.cpp
+++ b/src/conversions/cart.cpp
@@ -40,8 +40,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
-
 #include "proj_internal.h"
 #include <math.h>