--- a/src/conversions/geoc.cpp
+++ b/src/conversions/geoc.cpp
@@ -45,7 +45,7 @@ static void inverse(PJ_COORD &coo, PJ *P) {
     coo = pj_geocentric_latitude(P, PJ_INV, coo);
 }
 
-static PJ *CONVERSION(geoc, 1) {
+static PJ *PJ_CONVERSION(geoc, 1) {
     P->inv4d = inverse;
     P->fwd4d = forward;