--- a/src/conversions/geoc.cpp
+++ b/src/conversions/geoc.cpp
@@ -26,8 +26,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
-
 #include <math.h>
 
 #include "proj.h"