--- a/src/conversions/geocent.cpp
+++ b/src/conversions/geocent.cpp
@@ -27,8 +27,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
-
 #include "proj.h"
 #include "proj_internal.h"