--- a/src/conversions/noop.cpp
+++ b/src/conversions/noop.cpp
@@ -6,7 +6,7 @@ PROJ_HEAD(noop, "No operation");
 
 static void noop(PJ_COORD &, PJ *) {}
 
-PJ *CONVERSION(noop, 0) {
+PJ *PJ_CONVERSION(noop, 0) {
     P->fwd4d = noop;
     P->inv4d = noop;
     P->left = PJ_IO_UNITS_WHATEVER;