--- a/src/conversions/noop.cpp
+++ b/src/conversions/noop.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 
 #include "proj_internal.h"