--- a/src/conversions/set.cpp
+++ b/src/conversions/set.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 
 #include "proj_internal.h"
 #include <errno.h>