--- a/src/conversions/topocentric.cpp
+++ b/src/conversions/topocentric.cpp
@@ -25,8 +25,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
-
 #include "proj_internal.h"
 #include <errno.h>
 #include <math.h>