--- a/src/conversions/unitconvert.cpp
+++ b/src/conversions/unitconvert.cpp
@@ -63,8 +63,6 @@ Last update: 2017-05-16
 *
 ***********************************************************************/
 
-#define PJ_LIB_
-
 #include <errno.h>
 #include <math.h>
 #include <string.h>