--- a/src/factors.cpp
+++ b/src/factors.cpp
@@ -1,5 +1,5 @@
 /* projection scale factors */
-#define PJ_LIB_
+
 #include "proj.h"
 #include "proj_internal.h"
 #include <math.h>