--- a/src/init.cpp
+++ b/src/init.cpp
@@ -27,8 +27,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
-
 #include <ctype.h>
 #include <math.h>
 #include <stddef.h>