--- a/src/networkfilemanager.cpp
+++ b/src/networkfilemanager.cpp
@@ -1955,24 +1955,25 @@ NS_PROJ_END
 // ---------------------------------------------------------------------------
 
 #ifdef WIN32
-static const char dir_chars[] = "/\\";
+static const char nfm_dir_chars[] = "/\\";
 #else
-static const char dir_chars[] = "/";
+static const char nfm_dir_chars[] = "/";
 #endif
 
-static bool is_tilde_slash(const char *name) {
-    return *name == '~' && strchr(dir_chars, name[1]);
+static bool nfm_is_tilde_slash(const char *name) {
+    return *name == '~' && strchr(nfm_dir_chars, name[1]);
 }
 
-static bool is_rel_or_absolute_filename(const char *name) {
-    return strchr(dir_chars, *name) ||
-           (*name == '.' && strchr(dir_chars, name[1])) ||
-           (!strncmp(name, "..", 2) && strchr(dir_chars, name[2])) ||
-           (name[0] != '\0' && name[1] == ':' && strchr(dir_chars, name[2]));
+static bool nfm_is_rel_or_absolute_filename(const char *name) {
+    return strchr(nfm_dir_chars, *name) ||
+           (*name == '.' && strchr(nfm_dir_chars, name[1])) ||
+           (!strncmp(name, "..", 2) && strchr(nfm_dir_chars, name[2])) ||
+           (name[0] != '\0' && name[1] == ':' &&
+            strchr(nfm_dir_chars, name[2]));
 }
 
 static std::string build_url(PJ_CONTEXT *ctx, const char *name) {
-    if (!is_tilde_slash(name) && !is_rel_or_absolute_filename(name) &&
+    if (!nfm_is_tilde_slash(name) && !nfm_is_rel_or_absolute_filename(name) &&
         !starts_with(name, "http://") && !starts_with(name, "https://")) {
         std::string remote_file(proj_context_get_url_endpoint(ctx));
         if (!remote_file.empty()) {