--- a/src/pipeline.cpp
+++ b/src/pipeline.cpp
@@ -97,8 +97,6 @@ Thomas Knudsen, thokn@sdfe.dk, 2016-05-20
 *
 ********************************************************************************/
 
-#define PJ_LIB_
-
 #include <math.h>
 #include <stack>
 #include <stddef.h>