--- a/src/proj_mdist.cpp
+++ b/src/proj_mdist.cpp
@@ -27,7 +27,6 @@
 ** and inverse on unit ellipsoid.
 ** Precision commensurate with double precision.
 */
-#define PJ_LIB_
 
 #include <math.h>
 #include <stdlib.h>