--- a/src/projections/adams.cpp
+++ b/src/projections/adams.cpp
@@ -35,8 +35,6 @@
  *  https://en.wikipedia.org/wiki/Peirce_quincuncial_projection
  */
 
-#define PJ_LIB_
-
 #include <errno.h>
 #include <math.h>