--- a/src/projections/aea.cpp
+++ b/src/projections/aea.cpp
@@ -69,7 +69,7 @@ static double phi1_(double qs, double Te, double Tone_es) {
 }
 
 namespace { // anonymous namespace
-struct pj_opaque {
+struct pj_aea {
     double ec;
     double n;
     double c;
@@ -84,20 +84,20 @@ struct pj_opaque {
 };
 } // anonymous namespace
 
-static PJ *destructor(PJ *P, int errlev) { /* Destructor */
+static PJ *pj_aea_destructor(PJ *P, int errlev) { /* Destructor */
     if (nullptr == P)
         return nullptr;
 
     if (nullptr == P->opaque)
         return pj_default_destructor(P, errlev);
 
-    free(static_cast<struct pj_opaque *>(P->opaque)->en);
+    free(static_cast<struct pj_aea *>(P->opaque)->en);
     return pj_default_destructor(P, errlev);
 }
 
 static PJ_XY aea_e_forward(PJ_LP lp, PJ *P) { /* Ellipsoid/spheroid, forward */
     PJ_XY xy = {0.0, 0.0};
-    struct pj_opaque *Q = static_cast<struct pj_opaque *>(P->opaque);
+    struct pj_aea *Q = static_cast<struct pj_aea *>(P->opaque);
     Q->rho = Q->c - (Q->ellips ? Q->n * pj_qsfn(sin(lp.phi), P->e, P->one_es)
                                : Q->n2 * sin(lp.phi));
     if (Q->rho < 0.) {
@@ -113,7 +113,7 @@ static PJ_XY aea_e_forward(PJ_LP lp, PJ *P) { /* Ellipsoid/spheroid, forward */
 
 static PJ_LP aea_e_inverse(PJ_XY xy, PJ *P) { /* Ellipsoid/spheroid, inverse */
     PJ_LP lp = {0.0, 0.0};
-    struct pj_opaque *Q = static_cast<struct pj_opaque *>(P->opaque);
+    struct pj_aea *Q = static_cast<struct pj_aea *>(P->opaque);
     xy.y = Q->rho0 - xy.y;
     Q->rho = hypot(xy.x, xy.y);
     if (Q->rho != 0.0) {
@@ -155,7 +155,7 @@ static PJ_LP aea_e_inverse(PJ_XY xy, PJ *P) { /* Ellipsoid/spheroid, inverse */
 }
 
 static PJ *setup(PJ *P) {
-    struct pj_opaque *Q = static_cast<struct pj_opaque *>(P->opaque);
+    struct pj_aea *Q = static_cast<struct pj_aea *>(P->opaque);
 
     P->inv = aea_e_inverse;
     P->fwd = aea_e_forward;
@@ -163,17 +163,17 @@ static PJ *setup(PJ *P) {
     if (fabs(Q->phi1) > M_HALFPI) {
         proj_log_error(P,
                        _("Invalid value for lat_1: |lat_1| should be <= 90°"));
-        return destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
+        return pj_aea_destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
     }
     if (fabs(Q->phi2) > M_HALFPI) {
         proj_log_error(P,
                        _("Invalid value for lat_2: |lat_2| should be <= 90°"));
-        return destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
+        return pj_aea_destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
     }
     if (fabs(Q->phi1 + Q->phi2) < EPS10) {
         proj_log_error(P, _("Invalid value for lat_1 and lat_2: |lat_1 + "
                             "lat_2| should be > 0"));
-        return destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
+        return pj_aea_destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
     }
     double sinphi = sin(Q->phi1);
     Q->n = sinphi;
@@ -185,7 +185,7 @@ static PJ *setup(PJ *P) {
 
         Q->en = pj_enfn(P->n);
         if (Q->en == nullptr)
-            return destructor(P, 0);
+            return pj_aea_destructor(P, 0);
         m1 = pj_msfn(sinphi, cosphi, P->es);
         ml1 = pj_qsfn(sinphi, P->e, P->one_es);
         if (secant) { /* secant cone */
@@ -196,13 +196,14 @@ static PJ *setup(PJ *P) {
             m2 = pj_msfn(sinphi, cosphi, P->es);
             ml2 = pj_qsfn(sinphi, P->e, P->one_es);
             if (ml2 == ml1)
-                return destructor(P, 0);
+                return pj_aea_destructor(P, 0);
 
             Q->n = (m1 * m1 - m2 * m2) / (ml2 - ml1);
             if (Q->n == 0) {
                 // Not quite, but es is very close to 1...
                 proj_log_error(P, _("Invalid value for eccentricity"));
-                return destructor(P, PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
+                return pj_aea_destructor(P,
+                                         PROJ_ERR_INVALID_OP_ILLEGAL_ARG_VALUE);
             }
         }
         Q->ec = 1. - .5 * P->one_es * log((1. - P->e) / (1. + P->e)) / P->e;
@@ -222,28 +223,34 @@ static PJ *setup(PJ *P) {
     return P;
 }
 
-PJ *PROJECTION(aea) {
-    struct pj_opaque *Q =
-        static_cast<struct pj_opaque *>(calloc(1, sizeof(struct pj_opaque)));
+PJ *PJ_PROJECTION(aea) {
+    struct pj_aea *Q =
+        static_cast<struct pj_aea *>(calloc(1, sizeof(struct pj_aea)));
     if (nullptr == Q)
         return pj_default_destructor(P, PROJ_ERR_OTHER /*ENOMEM*/);
     P->opaque = Q;
-    P->destructor = destructor;
+    P->destructor = pj_aea_destructor;
 
     Q->phi1 = pj_param(P->ctx, P->params, "rlat_1").f;
     Q->phi2 = pj_param(P->ctx, P->params, "rlat_2").f;
     return setup(P);
 }
 
-PJ *PROJECTION(leac) {
-    struct pj_opaque *Q =
-        static_cast<struct pj_opaque *>(calloc(1, sizeof(struct pj_opaque)));
+PJ *PJ_PROJECTION(leac) {
+    struct pj_aea *Q =
+        static_cast<struct pj_aea *>(calloc(1, sizeof(struct pj_aea)));
     if (nullptr == Q)
         return pj_default_destructor(P, PROJ_ERR_OTHER /*ENOMEM*/);
     P->opaque = Q;
-    P->destructor = destructor;
+    P->destructor = pj_aea_destructor;
 
     Q->phi2 = pj_param(P->ctx, P->params, "rlat_1").f;
     Q->phi1 = pj_param(P->ctx, P->params, "bsouth").i ? -M_HALFPI : M_HALFPI;
     return setup(P);
 }
+
+#undef EPS10
+#undef TOL7
+#undef N_ITER
+#undef EPSILON
+#undef TOL