--- a/src/projections/airy.cpp
+++ b/src/projections/airy.cpp
@@ -26,7 +26,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-#define PJ_LIB_
 #include "proj.h"
 #include "proj_internal.h"
 #include <errno.h>