--- a/src/projections/august.cpp
+++ b/src/projections/august.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 
 #include <math.h>