--- a/src/projections/bacon.cpp
+++ b/src/projections/bacon.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 #include <errno.h>
 #include <math.h>