--- a/src/projections/bertin1953.cpp
+++ b/src/projections/bertin1953.cpp
@@ -9,8 +9,6 @@
   Port to PROJ by Philippe Rivière, 21 September 2018
 */
 
-#define PJ_LIB_
-
 #include <errno.h>
 #include <math.h>