--- a/src/projections/bipc.cpp
+++ b/src/projections/bipc.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 #include <errno.h>
 #include <math.h>