--- a/src/projections/boggs.cpp
+++ b/src/projections/boggs.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 #include <math.h>
 
 #include "proj.h"