--- a/src/projections/bonne.cpp
+++ b/src/projections/bonne.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 #include "proj.h"
 #include "proj_internal.h"
 #include <errno.h>