--- a/src/projections/calcofi.cpp
+++ b/src/projections/calcofi.cpp
@@ -1,4 +1,4 @@
-#define PJ_LIB_
+
 
 #include <math.h>