… .github/workflows/windows.yml
… docs/source/install.rst
… src/apps/CMakeLists.txt
… src/gauss.cpp
… src/projections/cass.cpp
… src/projections/cc.cpp
… src/projections/ccon.cpp
… src/projections/cea.cpp
… src/projections/chamb.cpp
… src/projections/col_urban.cpp
… src/projections/collg.cpp
… src/projections/comill.cpp
… src/projections/crast.cpp
… src/projections/denoy.cpp
… src/projections/eck1.cpp
… src/projections/eck2.cpp
… src/projections/eck3.cpp
… src/projections/eck4.cpp
… src/projections/eck5.cpp
… src/projections/eqc.cpp
… src/projections/eqdc.cpp
… src/projections/eqearth.cpp
… src/projections/fahey.cpp
… src/projections/fouc_s.cpp
… src/projections/gall.cpp
… src/projections/geos.cpp
… src/projections/gins8.cpp
… src/projections/gn_sinu.cpp
… src/projections/gnom.cpp
… src/projections/goode.cpp