Cf https://en.wikipedia.org/wiki/Unity_build and https://cmake.org/cmake/help/latest/prop_tgt/UNITY_BUILD.html

Timings on Linux with Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz, 12 vCPUS, and "make -j12 proj"

Before:
```
debug, unity disabled:
real    0m37,233s
user    5m28,169s
sys     0m28,826s

release, unity disabled:
real    0m45,684s
user    6m12,384s
sys     0m25,994s
```

After with  -DCMAKE_UNITY_BUILD=ON
```
debug, unity enabled:
real    0m19,870s
user    1m33,008s
sys     0m6,134s

release, unity enabled:
real    0m34,433s
user    2m20,730s
sys     0m5,395s
```

Debug builds with 12-threads are almost twice faster.
Release builds with 12-threads are 30% faster

And single-threaded release compilation in default mode takes 4m00s , and with CMAKE_UNITY_BUILD enabled only 1m50s