@jjimenezshaw 2023-11-23
This is a "risky" feature from CMake (we suffered it in my company).
I see that compiling time are better. But I personally feel (this is a personal opinion... or feeling) that it breaks some basic "law" in C/C++, the isolation of an object file from a single cpp file. In a way that may change if you add more files or whatever.
As you already experienced, it may not compile due to collisions among anonymous namespaces. They are not "anonymous" anymore.
It may happen the other way, that something builds with unity but not without. For instance if you forget an "#include" in a cpp file, that is in a previous one.

The error messages are not easy to understand... until you remember/realize that unity may be bothering you (that happened to me at work). And that happens in CI (until you enable unity in your local machine).

In this PR only some workflows in CI are moved to unity, so we could check how it behaves with and without. I do not know if you plan to move more platforms or configurations to unity.

If we include it we should document it existence and possible consequences in Contributing or Compiling documentation.

As far as I know, the only binary/package that we release are the docker files, right? We should compile then without unity.