@rouault 2023-11-23
Enabling unity build will not be done by default. The end user will have to enable it explictly if he wants to. We have a mix of non-unity and unity builds in the CI to be able to catch bug specifics to both type of configuration.

> If we include it we should document it existence and possible consequences in Contributing or Compiling documentation.

I've added the following note in install.rst:
```
.. option:: CMAKE_UNITY_BUILD=OFF

    .. versionadded:: 9.4

    Default is OFF. This can be set to ON to build PROJ using the
    https://cmake.org/cmake/help/latest/variable/CMAKE_UNITY_BUILD.html feature.
    This helps speeding PROJ build times. This feature is still considered
    experimental for now, and could hide subtle bugs (we are not aware of
    any at writing time though). We don't recommend it for mission critical
    builds.
```